+++ lib/storage_explorer/explorer_network_files.dart
@@ -8,10 +8,9 @@ import '../src/storage_database_exception.dart';
 import 'storage_explorer.dart';
 
 class ExplorerNetworkFiles {
-  final StorageExplorer storageExplorer;
   final ExplorerDirectory networkDirFiles;
 
-  ExplorerNetworkFiles(this.storageExplorer, this.networkDirFiles) {
+  ExplorerNetworkFiles(this.networkDirFiles) {
     _instance = this;
   }